+++ ChangeLog.md
@@ -3,6 +3,9 @@ ISAAC-rest Changelog
 This changelog summarizes changes and fixes which are a part of each revision.  For more details on the fixes, refer tracking numbers where provided, and the git commit history.  Note that this is not the same as the API Changelog.md.  This file will contain information on all changes - including bug fixes.  The API Changelog will only contain documentation on changes of the API - and those are tied to the 
 API version number, not the release version number.
 
+* 2017/08/?? - 4.25 - PENDING
+    *
+
 * 2017/08/24 - 4.24
     * Fix a bug with retirement or activation of extended description types on descriptions
     * Fix a bug with retirement of mapset sememes