+++ ChangeLog.md
@@ -3,6 +3,10 @@ ISAAC-rest Changelog
 This changelog summarizes changes and fixes which are a part of each revision.  For more details on the fixes, refer tracking numbers where provided, and the git commit history.  Note that this is not the same as the API Changelog.md.  This file will contain information on all changes - including bug fixes.  The API Changelog will only contain documentation on changes of the API - and those are tied to the 
 API version number, not the release version number.
 
+* 2017/08/31 - 5.01
+    * Adding a new system API which can be utilized to improve the module dropdowns in komet.
+    * Fix 572778 - change designation type and value set type to use EnglishLanguagePreferredTerm. 
+
 * 2017/08/25 - 5.00
     * ISAAC API alignment
     * Release 5 build